Alteração do ContaDao e adição do ContaDb e ContaRepository

+++ app/src/main/java/br/com/fiap/neoinbox/database/dao/ContaDao.kt
@@ -10,7 +10,7 @@ import br.com.fiap.neoinbox.model.Conta
 @Dao
 interface ContaDao {
     @Insert
-    fun salvarConta(conta: Conta): Int
+    fun salvarConta(conta: Conta): Long
 
     @Update
     fun atualizarConta(conta: Conta): Int
